Fixes for clang builds

+++ ref_gl/gl_dbghulls.cpp
@@ -45,7 +45,7 @@ void R_DrawWorldHull( void )
 
 	list_for_each_entry( poly, &hull->polys, chain )
 	{
-		srand((unsigned int)poly);
+		srand((unsigned int)(intptr_t)poly);
 		pglColor3f( rand() % 256 / 255.0, rand() % 256 / 255.0, rand() % 256 / 255.0 );
 		pglBegin( GL_POLYGON );
 		for( i = 0; i < poly->numpoints; i++ )
@@ -78,7 +78,7 @@ void R_DrawModelHull( void )
 	pglDisable( GL_TEXTURE_2D );
 	list_for_each_entry( poly, &hull->polys, chain )
 	{
-		srand((unsigned int)poly);
+		srand((unsigned int)(intptr_t)poly);
 		pglColor3f( rand() % 256 / 255.0, rand() % 256 / 255.0, rand() % 256 / 255.0 );
 		pglBegin( GL_POLYGON );
 		for( i = 0; i < poly->numpoints; i++ )
